requirements.txt
langchain
langchain-text-splitters
langchain-community
langchain-groq
langchain-huggingface
chromadb
pypdf
python-dotenv
